requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytz==2023.3
apscheduler==3.10.4
redis==5.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
ib-insync==0.9.86
yfinance==0.2.51

# Deployment and security
gunicorn==21.2.0
python-multipart==0.0.6
bcrypt==4.1.0
python-jose==3.3.0
